Rename route and controller to booking. Refactor booking to just have the controllers and bookingCalls to have the api calls
