+++ src/api/bookingCalls.js
@@ -1,24 +1,4 @@
 const axios = require("axios");
-const { StatusCodes } = require("http-status-codes");
-
-const getHotelsByLocation = async (req, res) => {
-  const { destination, checkinDate, checkoutDate, guestNumber } = req.body;
-
-  try {
-    const locationId = await searchLocation(destination);
-
-    const hotels = await searchHotels(
-      locationId,
-      checkinDate,
-      checkoutDate,
-      guestNumber
-    );
-
-    res.status(StatusCodes.OK).json(hotels);
-  } catch (error) {
-    res.status(StatusCodes.NOT_FOUND).json(error);
-  }
-};
 
 const searchLocation = async (destination) => {
   const options = {
@@ -107,4 +87,4 @@ const searchHotels = async (
   }
 };
 
-module.exports = { getHotelsByLocation };
+module.exports = { searchLocation, searchHotels };
